WAN 2.2 S2V: drop _wan suffix on files + consolidate loader/weightmap
File renames matching upstream wan/modules/s2v/ naming:
- motioner_wan.py → motioner.py (also delete MotionerTransformersWan
  dead-code placeholder that always raised NotImplementedError)
- audio_utils_wan.py → audio_utils.py (pure rename)

Consolidate the two paired S2V-only checkpoint helpers into a single
utils module:
- pipelines/wan/wan_s2v_loader.py + pipelines/wan/wan_s2v_weight_map.py
  → utils/wan_s2v_checkpoint.py

Both helpers are S2V-exclusive and have stdlib + torch deps only (no
pipeline coupling), so they belong under utils/ alongside other
checkpoint-handling code. Net: -362 lines / +13 lines + 1 new file.

Class names (FramePackMotionerWan, AudioInjector_WAN) and s2v_rope.py
left as-is — s2v_rope was investigated and has zero duplication with
existing WAN rope code; both functions are S2V-specific.

Verified: test_frame_packer_parity, test_prepare_audio_emb_parity,
test_audio_injector_parity (3 cases), test_prepare_rope_features_parity
all PASS after the rename. test_prepare_cond_emb_parity remains skipped
on the pre-existing binary_ng workaround.
